Finished setting up the links to work correctly, mostly by altering the bootstrap to allow the passing of paramaters. (ex.com/controller/method?paramId=paramValue&paramId=paramValue etc...)

+++ app/task/controllers/IndexController.php
@@ -22,24 +22,84 @@ class Task_IndexController extends Core_IndexController
     {
         $this->render();
     }
+
     public function add()
     {
-        $dbObj = new Task_DbDataMapperModel();
-        $dbObj->reorderTableIndex('Todo_List');
-        $dbObj->addTaskToTable($_POST['taskDescriptionTbox']);
+        if(isset($_POST['taskDescriptionTbox']))
+        {
+            $dbObj = new Task_DbDataMapperModel();
+            $dbObj->reorderTableIndex('Todo_List');
+            $dbObj->addTaskToTable($_POST['taskDescriptionTbox']);
+        }
+        else
+        {
+            //todo: throw error about the user not adding anything into the textbox.
+        }
         $this->render(__FUNCTION__);
     }
-    public function delete($taskToDelete)
+
+    public function delete($taskToDelete = null)
     {
-        var_dump($_SERVER);
-        $dbObj = new Task_DbDataMapperModel();
-        $dbObj->reorderTableIndex('Todo_List');
-        //you need to put the $_POST stuff here into an array...
-        $dbObj->deleteTasksFromTable($taskToDelete);
+        if($taskToDelete !== null)
+        {
+            if(count($taskToDelete) == 1)
+            {
+                foreach($taskToDelete as $key => $value)
+                {
+                    $taskToDelete = $value;
+                }
+
+                if($taskToDelete == '*')
+                {
+                    $dbObj = new Task_DbDataMapperModel();
+                    $dbObj->reorderTableIndex('Todo_List');
+                    $dbObj->deleteAllasksFromTable($taskToDelete);
+                }
+                else
+                {
+                    $dbObj = new Task_DbDataMapperModel();
+                    $dbObj->reorderTableIndex('Todo_List');
+                    $dbObj->deleteTasksFromTable($taskToDelete);
+                }
+            }
+            else
+            {
+                //todo: throw error about the user passing to many paramaters to this function...
+            }
+        }
+        else
+        {
+            //todo: throw error about the user not passing any information in to the method.
+        }
         $this->render(__FUNCTION__);
     }
-    public function update($itemToUpdate)
+
+    //todo: update query and this function...
+    public function update($taskToUpdate = null)
     {
+        if($taskToUpdate !== null)
+        {
+            if(count($taskToUpdate) == 1)
+            {
+                foreach($taskToUpdate as $key => $value)
+                {
+                    $taskToDelete = $value;
+                }
+                $dbObj = new Task_DbDataMapperModel();
+                $dbObj->reorderTableIndex('Todo_List');
+                //you need to put the $_POST stuff here into an array...
+                $dbObj->updateTasksFromTable($taskToUpdate);
+
+            }
+            else
+            {
+                //todo: throw error about the user passing to many parameters to this method...
+            }
+        }
+        else
+        {
+            //todo: throw an error about user not passing anything into the method.
+        }
         $this->render(__FUNCTION__);
     }
